helm: Add multiple improvements
This change includes:

- Allow the user to define which type of Kubernetes service to deploy, the
default continues to be `LoadBalancer`. However, we are now removing the
AWS LoadBalancer Controller annotations.

- Follow Kubernetes best practices in terms of memory resources & limits to
avoid Kubernetes OOM. Additionally, we are removing CPU limits to avoid CPU
throttling.

- Based on the above, we now control how much memory the readyset-server will
consume at all times. Therefore, we are now setting a memory limit of 95% for
the readyset-server through environmental variable `READYSET_MEMORY_LIMIT`
which was `0` before (Meaning it was able to consume all available memory).
The current default allows leaving free memory to be able to evict data
without OOMing.

- Reformat `values.yaml` file so services configurations are under
`readyset.adapter.service` and `readyset.server.service` keys.

- Change default database to `postgresql` (Previously it was `mysql`).

- Based on the `readyset.adapter.type` we set default ports for it (`5432`) or
(`3306`) if it is `mysql`.

- Leverage on `_helpers.tpl` for default settings so templates are cleaner.

- Add BATS unit tests for all mentioned changes, as well as configuring them
in Buildkite pipeline and bazel.

- Helm Chart formatting and add comments to `values.yaml` as well as `NOTES.TXT`

Change-Id: I4d5400e7d3d6f895985e5590bfa2190132924940
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5745
Reviewed-by: Luke Osborne <luke@readyset.io>
Tested-by: Buildkite CI