Update memory limit discription and name
Renamed `NORIA_MEMORY_BYTES` to `READYSET_MEMORY_LIMIT` to better convey
that this parameter is setting a limit on the memory usage. Also updated
the documentation strings provided by `--help`, which incorrectly said
that this was the amount of memory available for cached state rather
than the entire process.

Our public documentation on our website was already accurate for this
parameter and does not need to be updated.

Refs: REA-3148
Release-Note-Core: Renamed `NORIA_MEMORY_BYTES` to
  `READYSET_MEMORY_LIMIT` (and `--memory` to `--memory-limit,
  similarly), and updated the documentation provided by
  `--help` to be more accurate regarding the function of the parameter.
Change-Id: Ic71e09aa324f1d4755813cde179c9345d6abe506
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5646
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>