Adapter: Hide query-caching flag
Implicit and async caching modes are now considered experimental. This
is due to adjusted expectations on how we would like users to interact
with ReadySet rather than a regression.

The --query-caching flag is now hidden, and references to this
configuration in our docs are mostly removed. Remaining reference
specify that async and implicit caching modes are experimental.

Change-Id: I697a77d862ca61c1704a828fe98ab3294e05d11b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5654
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>