replicators: ACK PG log offset after writes
Postgres's replication protocol dictates that we periodically report the
LSN up to which we've successfully replicated in "standby status update"
messages. Previously, we only did this upon receipt of a keepalive
request message from the upstream database; however, there's no
guarantee that we'd receive a keepalive request message depending on the
settings of the upstream database and the replication stream throughput.
If we don't acknowledge our current position in the WAL relatively
frequently, the WAL file on the upstream database will grow without
bound.

This commit ensures that we report our log position roughly every ten
seconds.

Fixes: REA-3119
Release-Note-Core: Fixed an issue where we weren't reporting our
  position in the WAL during the replication of Postgres databases

Change-Id: I8dc33cada823e6409f7274d267cd21f007ca6515