readyset-psql: Simplify macro for types tests
This makes the macro a bit simpler by removing some unused features
(i.e. the ability to add arbitrary attributes to the test functions) and
changes the way we invoke the macro for clarity. It also standardizes
the way that we deal with strategies, which will make the following
commit a lot simpler when it comes time to test multiple values per test
invocation.

Refs: REA-3247
Change-Id: I2eb447cd139778f7be5cb897bc179ef2ac6e545d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5687
Tested-by: Buildkite CI
Reviewed-by: Aspen Smith <aspen@readyset.io>
Reviewed-by: Fran Noriega <fran@readyset.io>