readyset-psql: Improve types tests by testing more values
Most of the overhead in running each types test case is in the setup and
teardown: this is why we hardcode to only run 5 test cases per test. If
each test case tests a Vec of values though, then we can still limit
ourselves to 5 test cases while testing an order of magnitude more
values.

This slows the test down a little but, but with the sleep-removal
optimization I committed earlier, the tests run in about the same amount
of time that they used to.

Note that we define the builtin proptest `vec` strategy using a size
range because we want to be able to shrink the input down to a single
element. Ideally we could set the initial size of the Vec to always be
the same, and allow it to shrink down from there, but there doesn't seem
to be a way to guarantee we generate a big Vec at the start while still
allowing for shrinking. Running five test cases with a range from 1..20
seems like it should give us pretty decent coverage on average though
(and averages testing many more cases than we used to, when each test
case only generated a single value).

Fixes: REA-3247
Change-Id: I2c8856cc46af00ddfb13f8c8efe1d81d91d313e7