server: Remove conns for replicas that die
Within a running Replica, we maintain a map of established connections
for other domain replicas, to avoid having to re-establish a connection
on every message. If one of those replicas *died*, however, we'd never
remove the established connection from that map, meaning we'd never be
able to send messages to it if it ever came back.

Now, the `ChannelCoordinator` has a `tokio::sync::broadcast` channel
which can notify interested subscribers when the socket address for a
ReplicaAddress is changed - we subscribe to this in the Replica's
select-loop to remove the entry from that map of connections when we're
removing the configured socket from the channel coordinator.

This fixes an issue where after a worker died and came back, it'd never
be able to receive domain messages (such as full replay packets!).

This reverts commit 922f5028f2f75cc969805ec73033ceaf0f54320d.

Note that there is a *significant* performance issue here for large
deployments - there empirically seems to be a roughly quadratic slowdown
in `tokio::sync::broadcast::Sender::send` as the number of active
receivers increases - I've seen `send` take *minutes* when there are
more than like 100 receivers. We're band-aid-ing this issue (which is
why 922f5028f (Revert "server: Remove conns for replicas that die",
2023-08-03) happened in the first place) by avoiding sending anything on
the channel unless we *definitely* have to - but we ought to fix this at
some point in the future (it might be a tokio bug - I've reported it as
https://github.com/tokio-rs/tokio/issues/5923 and we'll see how that
goes).

Change-Id: Ic56e30a05e3b1e7552defca452c2fd5526692a55
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5680
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>