quickstart: Move download files into `/public/quickstart`
The compose files and seed data dumps were previously in `/public-docs`,
which was previously mirrored to a public github repo. This is no longer
the case, so move the quickstart files somewhere they will still be made
publicly available.

After merging this, we'll need to update the website to match:
https://github.com/readysettech/website-mono/pull/85

Change-Id: I4c8b02944563faca3a00053f0778993a4b7397a9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5699
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>