domain: Make tags in `ReplayPaths::by_dst` a HashSet
Now that as of e6f42f252 (server: Always add indexes during recovery,
2023-07-26) we're always adding indexes when planning out replay paths,
we can end up in a place where the message to setup a replay path in a
particular node gets sent more than once, eg if an ingress node is
shared by multiple downstream nodes. This is *mostly* fine, except that
due to the way the data structures work calling `ReplayPaths::insert`
wasn't properly idempotent if the same replay path was inserted more
than once. Having the same tag in this map multiple times would then
cause the same key to be replayed multiple times across the same replay
path, even though it had only been inserted into `self.waiting` once.
The first replay would then remove from `self.waiting`, causing the
second replay to *not* have its keys filtered out (since we don't do any
restricting of `for_keys` if nothing has been found in `self.waiting`
for the node). Then, when actually going to fill the holes for that
second replay, we'd (correcly!) panic in single_state due to trying to
fill an already-filled hole.

This commit fixes this by changing the deepest key type in
`ReplayPaths::by_dst` into a `HashSet` from a `Vec`, effectively making
`ReplayPaths::insert` idempotent. To avoid even getting to the place
where we try to fill an already-filled key, we also fail earlier with an
`internal` error if we get a replay piece targeting a node that is not
waiting for any keys at all.

This was found by system-benchmarks, but this commit covers it with a
logictest with two queries that setup nodes that share the same ingress,
which reproduced the original bug before the change to use a HashSet.

Fixes: REA-3261
Change-Id: I945326f2546875bf2133333725d8774f1df82d86
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5700
Reviewed-by: Luke Osborne <luke@readyset.io>
Tested-by: Buildkite CI