nom: Make sure exported parsers fail if there's remaining input
The `export_parser` macro is currently not checking that we are fully
parsing the bytes/strings that we use as input, so we end up ignoring
and discarding any remaining input.
This is a problem, since it leads to bugs like the one referenced by
this commit, where we parse Postgres type `int4range` as `int4`, which
then leads to us creating a table with incorrect column types.

This commit simply adds a check in the macro to make sure that there is
no remaining input, or fail the parsing otherwise.

Fixes: REA-3171
Change-Id: I579c1da6ab079ae9596dcfce6c9ff47a7a30caeb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5707
Tested-by: Buildkite CI
Reviewed-by: Aspen Smith <aspen@readyset.io>