replicators: Use UrlParseFailed error for URLs
Previously, we were using the `invalid_err!` macro when failing to parse
the upstream database URL in `NoriaAdapter`, which returned a
`ReadySetError::InvalidQuery` error. This resulted in an error message
that read "Invalid query" when, really, there was no query involved.

This commit replaces the `invalid_err!` invocations with
`ReadySetError::UrlParseFailed` errors when we fail to parse the
upstream database URL.

Change-Id: Iab1b46598ea29ce4461532c9958e57b03156f383