wip: dealing with local path madness
The docker-compose.yml in the root dir of the github repo is not the
same as the one used in the private repo.  I don't want to clobber the
current public one, so I copied the private repo one into ./build.
However, that meant the mounted workdir wasn't actually the root of the
repo.  This hack works for now, but we probably should standardize on a
single docker-compose.yml.