nom: Fixed expression parser failing when casting and using binary op
The `expression` parser was incorrectly parsing expressions such as
`'a'::abc < all('{b, c}')`, where it was leaving ` < all('{b, c}')` as
remaining input.
This occurred because when we parsed that expression, we first took the
value (in this case, "a") and then we attempted to either parse the
remaining input ("::abc < all('{b, c}')") as a binary op or a cast.

This was fixed by removing the `alt` (stopped checking for _either_ cast
 or binary op) and instead try to parse a cast first and then see if
 the remaining input corresponds to the binary op.

A test was added to check this scenario.

Refs: REA-3171
Change-Id: I9d60247911eab22e95a755e912680d341af01bc5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5719
Tested-by: Buildkite CI
Reviewed-by: Aspen Smith <aspen@readyset.io>