system-benchmarks: Fix config for simple benchmarks
The configuration for the 'simple' benchmark had a white space error
in the schema files, which prevented it from creating any rows in the
table. Further, the 'migrate' field was false, and thus the test was
not actually benchmarking readyset itself, but it's proxying behavior.
This was on both postgres and mysql.

Change-Id: I2af6565b9d29f5290c9c4a3b42e5c6415b33f99f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5720
Tested-by: Buildkite CI
Reviewed-by: Ethan Donowitz <ethan@readyset.io>