Updated README to work w/new URLs along w/changes to wording, sections, and flow
The README is being reworked to focus on getting people up
and running as quickly as possible. This means that some of the
content in the original README has been moved into 2 new files. One
of those files, next-steps, isn't used at this point but has
some important content in it. We'll most likely work it back into the
README via links.

Change-Id: If9a4e3dd067a072e00a6973bb425aae897b22a97
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5809
Reviewed-by: Luke Osborne <luke@readyset.io>
Tested-by: Buildkite CI