Replicators: Remove timeout from snapshot_completed()
Removes the timeout from TestChannel::snapshot_completed() which was
causing slow tests to fail. The timeout was added to avoid notifying the
controller that snapshotting had completed if the replicator exited with
an error, which was the previous behavior. Instead, that error is now
propagated and will cause the test to fail.

Fixes: REA-3335
Change-Id: Icdd88109b893026ea0d3103cd09f2260b26a0e26
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5767
Tested-by: Buildkite CI
Reviewed-by: Ethan Donowitz <ethan@readyset.io>