worker: Notify the controller about failed domains
If a domain fails due to a panic or error, send a message from the
worker to the new domain_died RPC on the controller to let it know that
the domain has died. This uses the previously-unused controller_uri in
the worker election state in the worker to obtain a (memoized) raw
controller handle.

Fixes: REA-2406
Change-Id: I6e8b7c2230c10eb9e32591c35eea784ebede9d7e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5768
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>