controller: remove_query return whether any expressions were removed
Modify the remove_query RPC to return 1 if one or more expressions were
removed, and 0 otherwise, and use this value to populate a DELETE result
from DROP CACHE. This feedback is useful e.g. if the user tries to drop
a non-existent cache name, or if the cache was already dropped
elsewhere.

Later, we may want to return the number of expressions (and aliases?)
removed.

Release-Note-Core: `DROP CACHE` now indicates whether a cache was
  removed via a DELETE result with a row count of either 1 or 0.

Change-Id: I620564a4146c72a4683e57ccd917ff45a651090f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5770
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>