errors: Attempt to unravel the tokio_tower::Error tangle
This commit makes a first attempt at trying to get actual useful errors
by downcasting rpc errors. I did a fair bit of poking around, and it
looks like the primary underlying error type we get if eg the server
side of something like a table rpc is dropped (for example, because the
domain has died) is a `tokio_tower::Error`. Downcasting to this to get
any actual useful data out of it is *remarkably* annoying since there
are a couple of associated types that it uses that're the actual
underlying stream/sink and request types for the transport, which took a
little finagling to get out (and pass to the rpc_err downcasting
function). That all works here (though this is tough to test
automatically).

To keep things clean and separated, I've also broken the `rpc_err`
stuff, including a placeholder sink type to pass in in the case where we
*don't* care about trying to downcast to a tokio_tower::Error.

Change-Id: I1b323d161022893b46f70478dde5b684dfb63d6e