dataflow-state: Add stateful proptest for weak indexes
This isn't quite complete yet, but it was surprisingly easy and
straightforward to write, and runs extremely quickly. The ultimate goal
is to test that the contents of weak indexes are correct, but for now
I've gotten the test to run through creating strict indexes and
inserting and removing rows without crashing, so I'm committing here as
a checkpoint.

Change-Id: I0fa52c3cd610c41c9043fbcbdbe1ec85b6b6ade5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5790
Reviewed-by: Aspen Smith <aspen@readyset.io>
Tested-by: Buildkite CI