dataflow-state: Add ReplayKey op to weak index proptest
This implements the last of the planned ops to the weak index stateful
proptest, allowing us to simulate the result of an upquery by filling a
hole and replaying any records that correspond to that key.

This immediately reproduced the bug described in #179, so I'm also
marking the test ignored for now.

Change-Id: I25c53901268e79094966a9b47d2a9cb723e89428
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5792
Tested-by: Buildkite CI
Reviewed-by: Aspen Smith <aspen@readyset.io>