clustertest: Use the right conn_str for newly added adapters
Use the correct connection string - with the right database type and
database name - for new adapters added with
DeploymentHandle::start_adapter, rather than hardcoding the mysql
connection string with the port of the adapter.

To work around REA-2787, this also db-qualifies table references in one
clustertest so that we can find tables to insert into.

Refs: REA-3356
Change-Id: I208607d78ff4d6390164cbb85ebfb8a7aecedc69
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5785
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>