dataflow-state: Add KeyedState::values method
This basically moves the existing SingleState::values method into
KeyedState, and then calls that from SingleState::values to retain the
same functionality in the existing method.

This could probably be justifiable as a refactor, since it avoids
SingleState needing to know as much about the implementation of the
KeyedState it wraps, but it also will be useful for testing purposes in
the following commit since it lets tests see what rows are contained in
a bare KeyedState without it needing to be wrapped in a SingleState.

Change-Id: I50cd0b2c1b4cda2f7e7934c0c7fbed1ac4a9a709
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5793
Tested-by: Buildkite CI
Reviewed-by: Aspen Smith <aspen@readyset.io>