dataflow-state: Add postcondition checks to weak index proptest
This fills out the last of the unimplemented code that I'd planned for
this test. With a fix in place in my local checkout for #179, this
immediately found a failing case for #286, just as I'd hoped.

Since it obviously still fails, I'm leaving it as ignored for now, and
we can revisit this once we have fixes in place for both of the
aforementioned issues.

Change-Id: I803664e2dd9d14f9e8ac7f75828f37dd83233cee