jepsen: Initial jepsen scaffolding to install readyset
Add a new directory containing the start of a Jepsen test for ReadySet.
Currently this (successfully!) bootstraps an HA ReadySet deployment,
with one server, a consul node, an upstream database, a load balancer,
and a configurable number of adapters, but does not run any tests.

Since I didn't want to reinvent the wheel, the stuff to set up consul is
either taken wholesale or adapted from the Jepsen test for consul
itself.

Coming soon: Actual tests, and some documentation on how to run this all
if you're not me.

Refs: REA-3357
Change-Id: Idd6a18d836130690ea8581724cbd3db514596a90