public: Fix postgresql transaction panic.
This corrects a panic triggered by a particular PostgreSQL transaction.
The transaction first added a new record to a table and then immediately
deleted all records from that same table. The panic was induced by
the remove operation on RocksDB. This operation checks if a
record exists in RocksDB before queuing the removal operation in the
RocksDB WriteBatch object. However, since the insertion and deletion
of a specific record were part of the same transaction, the record was
never actually written to RocksDB.

This change fixes it by removing records from the recordset that
would negate each other. This way there won't be a record that has
not been written to RocksDB that later needs to be removed.

Change-Id: I1270a3da00d352243116d62c237a4787c4d850f8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5801
Tested-by: Buildkite CI
Reviewed-by: Tamas Juhasz <tamas@readyset.io>
Reviewed-by: Luke Osborne <luke@readyset.io>