readyset: Allow configuring adapter external addrs
Add an `external_address` flag to the `readyset` binary, mirroring the
flag of the same name in the `readyset-server` binary, for configuring
the address that the adapter advertises to other readyset instances
running in the same deployment. This is important so that the server
knows how to connect to the reader domains running in the adapter for
embedded readers.

This also removes the `use_aws_external_address` flag from this binary,
since it was unused

Change-Id: I42793124923e9c66c2a5ec7b3d447fd4ba5e33c4