replicator: Persist last replicator error for user
Adds a `PersistentStats::last_replicator_error` which is updated when
the replicator hits an potentially recoverable error that causes it to
exit the main loop. These errors are propagated by splitting up the
`ReplicatorMessage::Error` variant into an `UnrecoverableError` and
`RecoverableError` variant.

If an error is logged, it will be show in `SHOW READYSET STATUS`. The
error is cleared after we enter the main replication loop as it
primarily exists to help users debug issues starting replication.

Release-Note-Core: Displays the last error encountered by the
  replicator in `SHOW READYSET STATUS` if the error prevents the
  replicator from entering the main replication loop. The error is
  cleared once replication is able to begin.
Change-Id: Iff0c77b37ba3ddba24811b495e6fa7d27a033d96