replicators: Use current COMMIT in PG offset
In 22368fcfe, we updated our definition of `PostgresPosition` to be an
ordered pair consisting of the LSN of the last commit we handled and the
LSN of an event in the next transaction. That definition did not account
for an edge case where, when beginning replication after completing a
snapshot, we mistakenly skipped the events in the first commit after the
snapshot's consistent point. This commit consists of data we haven't
seen yet, so skipping it meant data loss.

This commit updates the definition of `PostgresPosition` to be an
ordered pair consisting of the LSN of the commit that ends the *current*
transaction and the LSN of an event within that transaction. This
approach ensures that we correctly handle replication directly after
completing a snapshot.

Change-Id: Ib75811843ab85039d3d33fafb3f788a58359bfad