jepsen: Add a Client that reads from a table
Mostly as a proof-of-concept of actually performing operations on the
ReadySet cluster, add a jepsen Client that just reads all the data out
of a table (set-up by the client itself) repeatedly.

This involves generalizing a bunch of stuff out - there's a new
`nodes.clj` namespace containing all the node-type stuff, and most of
the code and constants for actually connecting to Readyset has been
moved into jepsen.readyset.client.

This works fine as of this commit, but obviously we're not doing
anything especially adversarial

Change-Id: I5d0114a0f9eae3c750d949b1ffb1037ad8bc909c