jepsen: Add a proof-of-concept Model for eventual consistency
Add a proof of concept of a very simple Model for a single eventually
consistent table. This writes rows to the table, and validates that the
results returned from reads always look like *one of* the valid results
from one of the table's past sets of rows.

This is all pretty hardcoded at this point, but going forward my plan is
to generalize it a bit to support multiple tables and arbitrary queries.

As of this commit, this all (probably unsurprisingly) still passes.

Change-Id: I8b542d4bb100fb716a735931e3be7518120fd64e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5817
Tested-by: Buildkite CI
Reviewed-by: Ethan Donowitz <ethan@readyset.io>