repl-offset: Make CommitLsn and Lsn opaque
`CommitLsn` and `Lsn` are types that represent Postgres LSNs and just
consist of a wrapped `i64`. Previously, the wrapped integer was made
public, allowing other modules and crates to get an integer back from
an LSN type returned by the Postgres WAL reader. Since `From<i64>` is
implemented for both `CommitLsn` and `Lsn`, this made it possible for
external modules/crates to convert an `Lsn` to a `CommitLsn` and vice
versa.

This commit makes the `CommitLsn` and `Lsn` types opaque, preventing
external modules/crates from retriving the wrapped integer. As part of
this process, this commit also moves the responsibility of converting a
string into a `CommitLsn` or `Lsn` to the `replication_offset::postgres`
module via `FromStr` implementations.

Change-Id: Ib235d4f610e423dffd6dfd3d139168a8003e8dd3