jepsen: Add the start of a simple liveness checker
Add the start of a checker for the ReadySet liveness guarantees. This
runs through the history keeping track of whether at least one adapter
is alive at any given time, and checks that all queries succeed as long
as at least one adapter is running.

Since we might kill an adapter while a query is actively running, this
also extends the Client to support a `:retry-queries?` option, which if
true will retry all queries once before considering them failed.

Change-Id: Ibda2d87f12e2f54a338cdd6884209ddce6b3ce0d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5833
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>