jepsen: Add a faster checker for commutative consistency
In the case that writes are commutative (which the ones we're testing
with currently are, and which we likely *always* will test with) we can
check eventual consistency much much more easily, by just running
through the ops saving the past valid results for tables directly,
without having to run through a linearizability checker. This shaves
a *ton* off of the asymptotic complexity of checking consistency, which
allows us to practically run *much* larger tests.

Refs: REA-69
Change-Id: Ic75d7af8298fa56699dfd9a881f775e20e4ad017
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5857
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>