database-utils: Provide more detail on parse failure.
If the upstream db url fails to parse due to special characters in the
password, it can result in confusing error messages like "Invalid port
number" because the parser doesn't fail until a later element in the
uri.

This commit adds more detail to the error message recommending that the
user percent-encode special characters as well as point to our
public documentation (since there may be other issues that cause the url
to fail parsing, and it's not easy to provide detailed guidance for
every possible case).

Release-Note-Core: Provided a better error message and guidance for
  passwords that include special characters, which must be
  percent-encoded.

Change-Id: I86a8f0ef3cc57eb9b1a9a0d956826c407002bc47