psql-srv: Rename PassThroughDataRow to avoid name clash
There are different types in tokio-postgres for data rows that are sent
as part of a SimpleQuery response vs data rows that are sent in response
to the parse/bind/describe/exec series of messages. We currently only
support passing through SimpleQuery rows, but we're about to add support
for doing the same thing for the other type of query. To minimize
confusion, I'm renaming the existing enum variant to make it clear it's
specific to SimpleQuery.

Refs: #268
Change-Id: I722f646e23cd6ae19ff12651e3cac9565e9221a6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5868
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>