psql-srv: Use SqlIdentifier for column names
Column names are usually short and are cloned often, so keeping them in
the stack rather than always using a heap reference (which is what
SqlIdentifier does) is likely much faster. This is an easy enough change
that I haven't benchmarked it, but I'd imagine it helps a bit especially
for very wide result sets.

Change-Id: Ia069100b48a21e5145b9555017cfe823471d4328
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5838
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>