mysql+coercion: Allow integers to be coerced to bits.
In MySQL, it's possible to have queries like these:
```
CREATE TABLE t (x bit(3));

SELECT * FROM t WHERE x = 7;
```
In this case, the value `7` will need to be coerced into a bit value.

Since MySQL treats bit values as unsigned ints[1], we try to
coerce the value into a `u64`.
From that number, using Rust's `Numeric` trait, we take the byte
representation from the value, and create a `BitVec`.

When then trim the `BitVec` to fit the column size (otherwise the size
would always be 64 bits), and check if that the discarded portion
contains only zeroes.
If the discarded part has a 1 in some position, then we are out of
bounds, and we fail.

This commit also adds a simple proptest that tries to coerce _any_
u64 values into a `DfType::Bit` + a MySQL adapter integration test to
account for a separate bug (REA-3381) found during the testing of this
fix.

[1]: https://dev.mysql.com/doc/refman/8.0/en/bit-value-literals.html

Refs: REA-3381
Fixes: REA-3300
Change-Id: I9600fb9af8207569c65fd7f685475b20377714d3