readyset-client: Add a `pg_meta` field to Change::CreateTable
As the first step of tracking the Postgres oids of tables and attrids of
their columns, add an optional (and currently unset and unused)
`pg_meta` field to `Change::CreateTable`. This will be passed by the
replicator, and saved on the Base node, so that we can return column
metadata to postgresql clients with table oids and column attribute IDs
filled in.

Refs: REA-3380
Change-Id: Ia5012b413d715bd62f6904a6bfda36d883ea670e