replicators: Pass table OID and column attnums in pg_meta
In both snapshotting and DDL replication, collect the oid of the table
and the attnums of its columns, and pass them as part of the new pg_meta
field when building `Change::CreateTable` changes to send to the
controller. We'll use these later when constructing result sets to fill
in the relevant fields on column metadata when columns are returned in
query results directly.

Refs: REA-3380
Change-Id: Ifa9b4bff5452e7f1ce1954f338aee48976b5ff75
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5844
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>