server: Store pg_meta on SqlIncorporator::base_schemas
Change the value type for `base_schemas` in SqlIncorporator from being
`CreateTableBody` directly to being a new `BaseSchema` struct, which
stores the `CreateTableBody` in addition to (now) the `pg_meta` for the
table. This is passed from the Change::CreateTable change in the recipe,
and will eventually be used when constructing postgres result set
metadata for queries that reference columns on tables directly.

Since a reference to `base_schemas` was previously passed directly to
the `RewriteContext` in `readyset_sql_passes`, some stuff had to change
there to make the various reference types and lifetimes work out.

Refs: REA-3380
Change-Id: I74903484a6cdc7138a5fb119c1054eb00428be6e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5848
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>