psql: Pass through pg_meta to result columns
If we know the PostgreSQL table OID and column attnum of a
column (because we know its ColumnBase, and it came from PG originally),
pass that all the way through to the `table_id` and `column_id` on the
FieldDescription we return in result sets to pg clients.

This is tough to test right know because tokio-postgres doesn't expose
these fields, but next up I'm going to patch it to do so, both so that
we can pass it through for proxied queries and so that we can test it
for cached queries. I have tested this manually using a Clojure REPL and
it does work. Before:

    user=> (jdbc/execute! ds ["select * from t1"])
    [{:a 1, :b nil}]

After:

    user=> (jdbc/execute! ds ["select * from t1"])
    [#:t1{:a 1, :b nil}]

Refs: REA-3380
Change-Id: Ide9867949212eea34ee1ec3575b53cb6f6fd4e66