jepsen: Generalize workloads + introduce votes
This commit significantly refactors the workloads (tables, rows
inserted, and queries run) that we use when testing ReadySet in Jepsen.
The client can now be configured with a list of tables to create and
queries to cache on init, and uses that same map of queries when
executing the `:query` op (renamed from `:read`) to generate SQL for
queries. The `commutative-eventual-consistency` checker now takes a map
from query ID to a function to compute the expected results for that
query given a map from table name to the list of rows for that table. To
demonstrate this all working, this comes along with a version of the
"votes" workload from the Noria thesis.

For representing queries (both SELECT statements and DDL) this uses the
HoneySQL library, which is nice as it allows us to both pass around and
introspect queries as Clojure data structures.

Since things are starting to get a *bit* complex, this starts writing a
few unit tests for the bits I wanted to make sure worked.

Fixes: REA-3386
Change-Id: I3c6eabf488b9bb287e83223f6824885ab8d7fbc5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5858
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>