replicatiors: Implemented --replication-tables-ignore
Sometimes we have an extensive number of tables but only a few of them
we want to avoid replication. Currently, we can only do this via
--replication-tables, However, if we want to filter out just one table,
we need to list all the ones we want, instead of the single one that we
do not want.

Added an option to blacklist the tables that are going to be
replicated.
This option is mutually exclusive with --replication-tables.

Release-Note-Core: Added a --replication-tables-ignore flag that allows
  you to replicate all tables other than the ones explicity ignored.
  Thanks, altmannmarcelo!

Change-Id: I916cf8f43e97c92f0264787f5dc7617a720832bd
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5865
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>