readyset-adapter: Add new type and param for passing metadata to exec
This adds the ability for different backends to pass extra
protocol-specific metadata to `execute`, which will then be passed on to
the upstream `execute` call if needed.

This new ability isn't used yet (both backends just set the type to
the `()` type) but it will be used in a followup commit to pass transfer
formats through for Postgres queries so that the same transfer formats
are used for proxied queries as for the original corresponding client
queries. This is necessary for enabling direct proxying of data rows for
query results so that the transfer format matches what the client
expects.

Refs: #268
Change-Id: Ic0ed365b96336e778ba34fabc6faea8c9abce404