jepsen: Generate deletes
Rename the previous `:insert` op to `:write` and generalize it to
support deletes in addition to just inserts, and make the generator for
writes (renamed from `:gen-insert` to `:gen-write`) also generate
deletes.

Since tracking the expected state of the db is getting complex and now
needs to happen in two places (the Writes generator to generate writes,
and the eventual-consistency checker to track the state of the db), this
also generalizes that out into a new `jepsen.readyset.memory-db`
namespace.

Fixes: REA-3396
Change-Id: Id523bc0d9482f1a09f0fa5b447375ad38ca11b40
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5866
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>