helm: Allow using an external Consul cluster
Until now, we only allowed the usage of the Consul cluster deployed from the
Consul Helm Chart which is a dependency of Readyset Helm Chart. Given the
Consul Server Service name is known beforehand the endpoint we were using
was hardcoded to `readyset-consul-server:8500`.

With this change, we allow the user to use an external Consul cluster by
disabling the deployment of the Consul Helm Chart with the value
`consul.enabled` set to `false` and `readyset.authority_address` configured
to the address of the Consul cluster.

By default we use the port `8500` if a port is not provided in the
`readyset.authority_address` setting.

Ref: REA-2855

Change-Id: Id0edfa5816c264800cd3875ff0ab87d35094fe27