jepsen: Generalize passing current rows to generators
Rather than packaging up the "track current rows and pass them to
generators" behavior in a custom Writes generator, make it into a
generator *wrapper* which can add tracking of the current rows within
the `context` map to *any* generator. This follows the pattern Jepsen
uses for its own generators more closely, and also allows us to use this
to generate more than just write ops that care about the current rows in
the table.

Change-Id: I5cf43b0de76c5a9ad042e373c8015aa523c560cf