readyset-psql: Finish implementation of raw row proxying for Execute
This ties together a bunch of previous commits to actually start passing
through raw rows in response to Postgres queries that use the extended
query protocol. We now:

 - Yield raw rows from the resultset Stream implementation instead of
   converting rows to Vec<PsqlValue>
 - Use the new ExecMeta feature to pass the list of transfer formats
   through to the upstream code
 - Use the new From impl from the last commit to convert TransferFormat
   to i16 before passing into the tokio-postgres code

It was necessary to start passing through the transfer formats in the
same commit where we start passing through the raw rows, since otherwise
the client may not get the result format it expects when it requests
text format for specific results.

Fixes: #268
Change-Id: I23b5554782d5be0d67c9b3652707129b4506ec1a