jepsen: Support parameterized queries
Extend our jepsen Client, workload abstraction, and eventual-consistency
checker to support parameterized queries in addition to queries without
params, to check eventual consistency of partial materialization. To
work around the fact that we now have one of many possible expected
results for a query at any point in time (one for each possible
parameter value!) the `:expected-results` fn in a workload can now
return a *function* taking params instead of a vector of expected
results - we track these functions in a list alongside the set of
vectors of expected results, and call each function on the current set
of params for each partial query. This might get kinda slow at some
point, but feels fine for now.

Fixes: REA-3397
Change-Id: If14cbffc2e9152915bac9135d489d46e7955afb4