replicators: Return table error when inserting unsupported type.
We were returning a `WalError::ReadySetError` when we were unable to
parse the type of a column when processing an insert during replication
(something that doesn't happen during snapshot).
As a result of that, replication would fail, make the adapter restart,
only to keep encountering the same insert and failing over and over
again.

To fix that, this commit simply changes the error being returned to be
`WalError::TableError`.
This will correctly mark the table as unsupported.

Fixes: REA-3170
Change-Id: I64e513263fda1e056005ad3e04b34eafa66e6785