nom-sql: Added support for SHOW PROXIED SUPPORTED QUERIES
On a heavy-load system, SHOW PROXIED QUERIES can show a large number
of queries that might not be eligible for caching.

Added support for a variant of SHOW PROXIED QUERIES that accept an
optional [SUPPORTED] keyword. If present, only queries that are
eligible for caching are shown.

Release-Note-Core: You can now optionally display only supported proxied
  queries with the command `SHOW PROXIED SUPPORTED QUERIES`. Thanks,
  altmannmarcelo!
Change-Id: I9611ae07c3fb50e477ac198d7e86f161000de1cb