readyset-psql: Fix repl/resnapshot test
This commit reworks a test that was added to test that we correctly
replicate writes that are sent to the upstream while we are taking a
snapshot. Previously, the test relied on a lot of sleeping, making it
slow and prone to flakiness. Now, it uses the `eventually!` macro to
poll for the correct results.

Change-Id: I5d1a59ddb831bec35ad6268b0e423e91456b4202
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5915
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>