readyset-util: Allow `eventually` to return a result
If the last (eg a non-panicking) invocation of the `then_assert` block
in an invocation of `eventually` yields a result, make that the result
of the *overall* invocation of the `eventually` macro. This allows us to
keep trying to obtain some value within a test using `eventually`, in
addition to just asserting on some condition.

Change-Id: I6ceee248c74de3977473310061280b8e4e4576d2