readyset-psql: Wait for tables to replicate in test
The `show_proxied_queries_show_caches_query_text_matches` test was
creating a table, then *immediately* querying from that table and
expecting a cache to be created for that query. This could frequently
flake in CI if the table hadn't been replicated yet. Instead, this
commit wraps the bit that tries to run the query and look up its cache
name in an `eventually`, to allow for it to happen more slowly.

Change-Id: I8796244262917a103700a02c847c040675115973